react-gbajs: refactor improving code and folder structure
