+++ react-gbajs/COPYING
