+++ react-gbajs/src/emulator/arm.js
