+++ react-gbajs/src/emulator/audio.js
