+++ react-gbajs/src/emulator/core.js
