+++ react-gbajs/src/emulator/gba.js
