+++ react-gbajs/src/emulator/io.js
