+++ react-gbajs/src/emulator/irq.js
