+++ react-gbajs/src/emulator/savedata.js
