+++ react-gbajs/src/emulator/setVolume.js
@@ -1,0 +1,4 @@
+const setVolume = (gba, value) =>
+  gba.audio.masterVolume = Math.pow(2, value) - 1
+
+export default setVolume
