+++ react-gbajs/src/emulator/sio.js
