+++ react-gbajs/src/emulator/thumb.js
