+++ react-gbajs/src/emulator/util.js
