+++ react-gbajs/src/emulator/video.js
