+++ react-gbajs/src/emulator/video/proxy.js
