+++ react-gbajs/src/emulator/video/software.js
