+++ react-gbajs/src/emulator/video/worker.js
