--- react-gbajs/src/resources/xhr.js
@@ -1,8 +1,0 @@
-window.loadRom = function loadRom(url, callback) {
-	var xhr = new XMLHttpRequest();
-	xhr.open('GET', url);
-	xhr.responseType = 'arraybuffer';
-
-	xhr.onload = function() { callback(xhr.response) };
-	xhr.send();
-}
